+++ app/src/main/java/com/jmarser/weatherapp_java/forecast/presenter/ForecastPresenter.java
@@ -2,5 +2,5 @@ package com.jmarser.weatherapp_java.forecast.presenter;
 
 public interface ForecastPresenter {
 
-    void getForecast(Long lat, Long lon);
+    void getForecast(double lat, double lon);
 }
